refactor(web): consolidate Go files and centralize models
